+++ client/src/pages/user/UserRow.js
@@ -56,114 +56,7 @@ class GroupCard extends Component {
 			style.backgroundColor = "#e0e0e0";
 			let sortValue = this.props.sortValue;
 			let sortDirection = this.props.sortDirection === "asc" ? "asc" : "desc";
-			console.log(sortDirection);
-			return (
-			<tr>
-				<Paper
-					elevation={this.state.hover ? 2 : 1}
-					onMouseEnter={() => this.setState({ hover: true })}
-					onMouseLeave={() => this.setState({ hover: false })}
-					component="td"
-					style={style}
-				>
-					<div style={{
-						display: "flex",
-						justifyContent: "space-between"
-					}}>
-						<Typography variant="title" color="primary" style={{ flex: 1}} >
-							<Tooltip title="Sorteer" enterDelay={300} placement="bottom-start">
-                <TableSortLabel active={sortValue === "name"}
-                  direction={sortDirection}
-                  onClick={() => this.props.onSortChange("name")}
-                  style={{color:"inherit"}} >
-                  Naam
-                </TableSortLabel>
-              </Tooltip>
-						</Typography>
-						<Typography variant="subheading" style={{ flex: 1 }} >
-              <Tooltip title="Sorteer" enterDelay={300} placement="bottom-start">
-                <TableSortLabel active={sortValue === "school"}
-                  direction={sortDirection}
-                  onClick={() => this.props.onSortChange("school")} >
-                  School 
-                </TableSortLabel>
-              </Tooltip>
-						</Typography>
-						<Typography variant="body1" style={{ flex: 1 }} >
-              <Tooltip title="Sorteer" enterDelay={300} placement="bottom-start">
-                <TableSortLabel active={sortValue === "levelAndYear"}
-                  direction={sortDirection}
-                  onClick={() => this.props.onSortChange("levelAndYear")} >
-                  Niveau - Leerjaar 
-                </TableSortLabel>
-              </Tooltip>
-						</Typography>
-						<Typography variant="body1" style={{ flex: 1 }} >
-              <Tooltip title="Sorteer" enterDelay={300} placement="bottom-start">
-                <TableSortLabel active={sortValue === "role"}
-                  direction={sortDirection}
-                  onClick={() => this.props.onSortChange("role")} >
-                  Rol
-                </TableSortLabel>
-              </Tooltip>
-						</Typography>
-						<Typography variant="body1" style={{ flex: 1 }} >
-              <Tooltip title="Sorteer" enterDelay={300} placement="bottom-start">
-                <TableSortLabel active={sortValue === "profile"}
-                  direction={sortDirection}
-                  onClick={() => this.props.onSortChange("profile")} >
-                  Profiel 
-                </TableSortLabel>
-              </Tooltip>
-						</Typography>
-					</div>
-					{this.props.role === "admin" &&
-						<div style={{
-							display: "flex",
-							justifyContent: "space-between"
-						}}>
-							<Typography variant="body1" style={{ flex: 1 }} >
-                <Tooltip title="Sorteer" enterDelay={300} placement="bottom-start">
-                  <TableSortLabel active={sortValue === "email"}
-                    direction={sortDirection}
-                    onClick={() => this.props.onSortChange("email")} >
-                    Office Email
-                  </TableSortLabel>
-                </Tooltip>
-							</Typography>
-							<Typography variant="body1" style={{ flex: 1 }} >
-                <Tooltip title="Sorteer" enterDelay={300} placement="bottom-start">
-                  <TableSortLabel active={sortValue === "preferedEmail"}
-                    direction={sortDirection}
-                    onClick={() => this.props.onSortChange("preferedEmail")} >
-                    Voorkeurs email 
-                  </TableSortLabel>
-                </Tooltip>
-							</Typography>
-							<div style={{ flex: 1 }} />
-							<Typography variant="body1" style={{ flex: 1 }} >
-                <Tooltip title="Sorteer" enterDelay={300} placement="bottom-start">
-                  <TableSortLabel active={sortValue === "phoneNumber"}
-                    direction={sortDirection}
-                    onClick={() => this.props.onSortChange("phoneNumber")} >
-                    Telefoonnummer
-                  </TableSortLabel>
-                </Tooltip>
-							</Typography>
-							<Typography variant="body1" style={{ flex: 1 }} >
-                <Tooltip title="Sorteer" enterDelay={300} placement="bottom-start">
-                  <TableSortLabel active={sortValue === "id"}
-                    direction={sortDirection}
-                    onClick={() => this.props.onSortChange("id")} >
-                    Gebruikers ID
-                  </TableSortLabel>
-                </Tooltip>
-							</Typography>
-						</div>
-					}
-				</Paper >
-			</tr>
-			)
+			console.log(this.props);
 		}
 		return (
 			<tr>
